+++ dedoc/data_structures/concrete_annotations/reference_annotation.py
@@ -27,7 +27,7 @@ class ReferenceAnnotation(Annotation):
                     page_id=10,
                     line_id=189,
                     tag_hierarchy_level=HierarchyLevel(level1=2, level2=0, paragraph_type="bibliography_item")),
-                    other_fields={"uid": "97cfac39-f0e3-11ee-b81c-b88584b4e4a1"}
+                    uid="97cfac39-f0e3-11ee-b81c-b88584b4e4a1"
                 ),
                 annotations=[]
             )